rename mxfp -> mxfp_simple
