+++ src/mase_triton/mxfp_simple/functional/__init__.py
