+++ src/mase_triton/mxfp_simple/functional/cast.py
