+++ src/mase_triton/mxfp_simple/functional/linear.py
